Added new container classes in container.py which are high-level.
Much like the BarContainer, the HistContainer and ScatterContainer
classes are meant to save some of the high-level information about the
histogram and scatter plots that a user is attempting to draw. These are
attached to the axes.containers list. These can then be accessed later
with the references to the actual draw-able objects included.